Move `doubleClick.ts` to `@/util`
