+++ app/gui2/src/components/GraphEditor/GraphNode.vue
@@ -5,7 +5,6 @@ import GraphNodeError from '@/components/GraphEditor/GraphNodeError.vue'
 import GraphVisualization from '@/components/GraphEditor/GraphVisualization.vue'
 import NodeWidgetTree from '@/components/GraphEditor/NodeWidgetTree.vue'
 import SvgIcon from '@/components/SvgIcon.vue'
-import { useDoubleClick } from '@/composables/doubleClick'
 import { injectGraphNavigator } from '@/providers/graphNavigator'
 import { injectGraphSelection } from '@/providers/graphSelection'
 import { useGraphStore, type Node } from '@/stores/graph'
@@ -15,6 +14,7 @@ import { Rect } from '@/util/data/rect'
 import { Vec2 } from '@/util/data/vec2'
 import { displayedIconOf } from '@/util/getIconName'
 import { useApproach } from '@/util/vue/animation'
+import { useDoubleClick } from '@/util/vue/doubleClick'
 import { usePointer, useResizeObserver } from '@/util/vue/events'
 import { setIfUndefined } from 'lib0/map'
 import type { ContentRange, ExprId, VisualizationIdentifier } from 'shared/yjsModel'
